Simplify demo player
